Change abbreviation characters
While ... is valid in Windows, .NET does not recognize this as a valid path.  As a result, tools like msbuild will fail.  After abbreviation, change ... to --- to avoid issues

Also fix tests on Windows
Windows has \ as the path separator character, and a significantly shorter MAX_PATH.  Set the cached path length long enough that a test that doesn't want the path to be shortened passes.
Also replace / with \ on Windows (and vice versa for non-Windows platforms) when comparing paths.